requirements.txt
Flask
flask-cors
requests
pymongo
python-dotenv
werkzeug
google-generativeai
joblib
PyPDF2
python-docx
numpy
gensim
tensorflow
pickle-mixin
pandas
nltk
Flask
Flask-Cors
PyMuPDF
requests
textract
google-cloud-generative-ai
joblib
PyPDF2
python-docx
numpy
gensim
tensorflow
pickle
pandas
nltk
